Return ticket function was added
