+++ src/main/java/com/epam/aviasales/services/ProfileService.java
@@ -1,31 +1,10 @@
 package com.epam.aviasales.services;
 
 import com.epam.aviasales.domain.Ticket;
-import com.epam.aviasales.repositories.TicketRepository;
-import com.epam.aviasales.repositories.impl.TicketRepositoryImpl;
-
 import java.util.List;
 
-public class ProfileService {
-  private static volatile ProfileService instance;
-  private static final TicketRepository ticketRepository = TicketRepositoryImpl.getInstance();
+public interface ProfileService {
+  List<Ticket> getAccountTickets(Long accountId);
+  void deleteAccountTicketById(Long ticketId);
 
-  private ProfileService() {}
-
-  public static ProfileService getInstance() {
-    ProfileService localInstance = instance;
-    if (localInstance == null) {
-      synchronized (ProfileService.class) {
-        localInstance = instance;
-        if (localInstance == null) {
-          instance = localInstance = new ProfileService();
-        }
-      }
-    }
-    return localInstance;
-  }
-
-  public List<Ticket> getAccountTickets(Long accountId) {
-    return ticketRepository.getTicketsByAccountId(accountId);
-  }
 }
